requirements.txt
discord.py
python-dotenv
animec
requests